requirements.txt
feedparser==6.0.11
requests==2.31.0
openai>=1.55.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
